♻️ Refactor to use routes folder
